# internal/oas/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn3AllowedHeaders = map[string]string{
		"POST": "Content-Type,X-Swarm-Deploy-Secret",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/v1/"

			if l := len("/api/v1/"); len(elem) >= l && elem[0:l] == "/api/v1/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 's': // Prefix: "s"

				if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 't': // Prefix: "tacks"

					if l := len("tacks"); len(elem) >= l && elem[0:l] == "tacks" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleListStacksRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'y': // Prefix: "ync"

					if l := len("ync"); len(elem) >= l && elem[0:l] == "ync" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleTriggerSyncRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'w': // Prefix: "webhooks/git"

				if l := len("webhooks/git"); len(elem) >= l && elem[0:l] == "webhooks/git" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleReceiveGitWebhookRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/v1/"

			if l := len("/api/v1/"); len(elem) >= l && elem[0:l] == "/api/v1/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 's': // Prefix: "s"

				if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 't': // Prefix: "tacks"

					if l := len("tacks"); len(elem) >= l && elem[0:l] == "tacks" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = ListStacksOperation
							r.summary = ""
							r.operationID = "listStacks"
							r.operationGroup = ""
							r.pathPattern = "/api/v1/stacks"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'y': // Prefix: "ync"

					if l := len("ync"); len(elem) >= l && elem[0:l] == "ync" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = TriggerSyncOperation
							r.summary = ""
							r.operationID = "triggerSync"
							r.operationGroup = ""
							r.pathPattern = "/api/v1/sync"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'w': // Prefix: "webhooks/git"

				if l := len("webhooks/git"); len(elem) >= l && elem[0:l] == "webhooks/git" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = ReceiveGitWebhookOperation
						r.summary = ""
						r.operationID = "receiveGitWebhook"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/webhooks/git"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}